内存计算使用double类型，避免因精度偏差过大

+++ src/main/java/fit/lang/plugin/json/monitor/GetMonitorDataJsonExecuteNode.java
@@ -28,7 +28,7 @@ public class GetMonitorDataJsonExecuteNode extends JsonExecuteNode {
         result.put("second", second);
 
         int cpuCount = JsonExecuteNodeMonitorUtil.getCpuProcessorCount();
-        long memoryG = getMemoryG();
+        double memoryG = getMemoryG();
         result.put("cpuCount", cpuCount);
         result.put("memoryG", memoryG);
